Added libs/common/ngx/environment
Contains environment service which handles the availability of the environment DI token value instead of delegating that responsibility to any other service/component that requires environment variables.

Currently contains a single "getter" function that tests and returns the value of an environment variable, if it exists. Accepts a string reference, but this should be done with symbols down the road if possibe to reduce the probability of trying to access an invalid variable from an incorrect spelling