Fixed UES mobile map dragging
Map was behaving unexpectedly when the map was dragged on a mobile view. Almost seemed like something was rejecting touch inputs.

Rolled back esri-loader dependency to load 4.12 esri modules, instead of 4.15. NPM package json was incorrect and was installing the latest minor patch, which was loading 4.15. Applications have not been tested on 4.15 so rolling back to 4.12 was necessary. The mobile issue was due to a mismatch beteween esri CSS styles and module versions.

Added generators as a search source