UES Integration (#59)
* Added base app and lib

* Reconciled old component refs vs those in the repo

* Pulled out aggiemap desktop sidebar into lib

* Transferred over some mobile ui components

- MobileUI Base component
- Trip planner top component
- Trip planner bottom component
- Omnisearch component

* Moved over the remaining mobile components

- Sidebar
- Sidebar TOC
- Modal
- Form components

* Pulled out the frame/skeleton module

* Creating UES specific components and normalizing configs

* Categorized layer list now uses accordion component

* Keeping layer list expanded on layer model change

The layer model changes whenever there are property changes to individual layers (visible vs not) which causes the layer list to re-render. This change ensures that the layer list remains expanded if it was previously expanded, otherwise it would re-render all categories as collapsed.

* Updated layer and popup definitions

* Updated fire hydrants search source

* Added UES branding blocks

* Fixed routing issue

* Fixed construction layer popup data fields

* Fixed UES mobile map dragging

Map was behaving unexpectedly when the map was dragged on a mobile view. Almost seemed like something was rejecting touch inputs.

Rolled back esri-loader dependency to load 4.12 esri modules, instead of 4.15. NPM package json was incorrect and was installing the latest minor patch, which was loading 4.15. Applications have not been tested on 4.15 so rolling back to 4.12 was necessary. The mobile issue was due to a mismatch beteween esri CSS styles and module versions.

Added generators as a search source

* Added more active search sources

* Fixed imports