dep(nx-workspace): Nx workspace upgrade (#90)
* Updated to workspace 8.12.7 and generated migrations

This was a manual upgrade to 8.12.7 from 8.8.0. Migrations from and to have been generated but not been executed

* Executed migrations from  8.8.0 => 8.12.7

* Upgraded from 8.12.7 => 8.12.10

No migrations generated with this upgrade so nothing additional to run

* Generated migrations 8.12.10 => 9.0.0

* Manually updated @nrwl/cli to 9.0.0 due to failed migration

Unregistered task "run-schematic" in schematic "update-9-0-0".

* Executed migrations to 9.0.0 from 8.12.10

One migration failed

* Generated migrations from 9.0.0 to 9.8.0

* Executed migrations to 9.8.0

Failed migrations:

@nrwl/cypress - ./src/migrations/update-9-4-0/add-cypress-eslint-plugin-9-4-0
@nrwl/web - ./src/migrations/update-9-2-0/set-build-libs-from-source

* Generated migrations from 9.8.0 to 10.0.0

* Executed migrations for 10.0.0

Had some false-positive errors with https://v10.angular.io/guide/migration-undecorated-classes with some service classes but all decorators appear to be correct

* Generated migrations form 10.0.0 to 10.4.7

* Executed migrations to 10.4.7 from 10.0.0

* Generated migrations from 10.4.7 to 11.0.0

* Executed migrations for 11.0.0

* Generated migrations from 11.0.0 to 11.0.16

* Executed migrations for 11.0.16

* Changed and updated angular-webstorage-service => ngx-webstorage-service

* Fixed obvious bug in aggiemap-angular with bus route list

Bus route component had a forwardRef applied and was causing issue registering ngOnDestroy.

* Added arcgis-js-api type to app tsconfigs

* Converted to angular.json to workspace.json

* nx:format pass

* Changed projects VSC TS version

Uses whatever is the installed workspace version. As of a recent VSC update, it was defaulting to VSC's default version and causing library execution "errors"

* Moved @nrwl/angular to dev dependencies