Fixed issue where map service singleton being overwritten on map component destroy/re-initliazation.
Everytime the map component was destroyed due to a route change, or otherwise, on  component re-initialization, a completely new map service was being created but the previous one was not destroyed and any handles and refereces that other components depended on simply went unresponsive. This made it impossible for things like layer lists, legends, etc, to re-render their contents on re-initialization.