refactor(maps-feature-legend): Esri legend component rewrite (#102)
* Rendering legend items from legend view model

Need to decide whether we want to nuke the old legend logic which dervies legend items from manualy added legend sources  in the app envirionments, or we want to assimilate it into the current which would give us another way to override symbology even though the rednerer property on a feature layer does just that already.

* Moved legend logic to service to maintain active layer state

State was being reset when the logic lived in the component, instead of the service.

* Fixed issue where map service singleton being overwritten on map component destroy/re-initliazation.

Everytime the map component was destroyed due to a route change, or otherwise, on  component re-initialization, a completely new map service was being created but the previous one was not destroyed and any handles and refereces that other components depended on simply went unresponsive. This made it impossible for things like layer lists, legends, etc, to re-render their contents on re-initialization.

* Temp fix for environment side-effects from mutability

Primarily affecting `LayerSources`, where the esri map service is directly updating the value and causing undesired behavior (layer sources persisting in memory) from esri map service creation that affects the initial loading conditions of map re-instantiation.

Since there is some rework scheduled for the layer lists, which currently depend on environment LayerSources, a more appropriate solution will be implemented for this.

* Restructured legend for better presentation

Collapsing one-item legend groups.

* Removed layer list component change

Change is outside of scope of this feature branch.