Fixed popups not resolving due to layer service changes
Since layer state is no longer maintained by me, basically everything was using layer sources environment for lookup and since that duty has been delegated to the layer list view model, the added layers to the map did not exist in layer sources, and thus the popup references were not available.

The popup references have been appended to the layer prototype, so they should be available for everything there too.