Added support for static legend items
These are items that are not added on to the map as a layer and thus cannot have a legend info rendered for it, but need to represented in the legend.

Aggiemap has some of these because the entire map server is not used as a source and certain elements are not treated as features and ae not interactible.

Updated aggiemap legend infos