Added scenario re-ordering
Fixed scenario builder race condition on scenario creation. On successful creation, the routing to the scenario details would happen prematulrely leading to a blank form in some cases.

Changed the way the scenario adding and removing works. Used to immediately issue a request to create/remove the entry but now requires a user form submission to trigger the requests.