Added conditional affected base
If the build source branch is development or master, it will set the base to be one commit before the head that way the affected command can compute affected apps/libs between old and new commit. This solves the base and head both being equal to each other and not calculating any changes.