Setup conditional affected base for nx commands
Removed development state condition that stated the stage would only run if the build source branch name was not development or master. This meant that merged PR's would not trigger the build pipeline.

Instead, using an affectedBase variable that derives it's value based on the build source branch name so that the affected commands have the correct base and target values so to fix the original issue that prompted the initial above change because head and target were equal on commits on development.