ci: Added  conditional affected base variable for nx commands (#125)
* Set builds to run on anything but dev or master

Condition on development stage has been set so that the stage will only trigger if a build source branch is anything other than dev or master. This is because the current affected command relies on base HEAD and target development. On the development branch base and target are the same, meaning that no artifacts are produced which will consequently fail to trigger a release pipeline.

Added production stage dependency on development since if development does not build, neither should production. Both use the same build options, just different configs.

Added tag assignment stage depedency on development and production since both  must pass for any release trigger.

* Added conditional affected base

If the build source branch is development or master, it will set the base to be one commit before the head that way the affected command can compute affected apps/libs between old and new commit. This solves the base and head both being equal to each other and not calculating any changes.

* Setup conditional affected base for nx commands

Removed development state condition that stated the stage would only run if the build source branch name was not development or master. This meant that merged PR's would not trigger the build pipeline.

Instead, using an affectedBase variable that derives it's value based on the build source branch name so that the affected commands have the correct base and target values so to fix the original issue that prompted the initial above change because head and target were equal on commits on development.