maint(oidc): Scope cleanup for the OIDC apps/libraries (#123)
* Removed oidc-provider-angular

This was a boilerplate application that was supposed to be server-side rendered and work along with oidc-provider-express. However, that project failed.

* Removed oidc-provider-express

This project was the express backend for oidc-provider-angular. This project failed and was superceded by oidc-provider-nest

* Removed oidc-provider library

This library worked alongside oidc-provider-express which has been removed.

* Sneaking a library tag fix

Unsure if the extra space matters or not, but at least it's consistent with other tag formats.