Reverted map-image layer opacity control
The control has handed off to the children individually but went back to let the parent handle opacity for all children. I experienced and issue with that behavior before but it caused a bug where trying to apply an opacity to sub-layers that was higher than they were currently set did not apply the changes to the map raster itself until the page was reloaded.

Example:

Initial opacity for all sub-layers = .5;

Setting the opacity to 1 on all sub-layers did not change the raster on the map. However, upon refreshing the page, the change to opacity 1 was visible.