feat(cpa-angular): Map image layer support (#126)
* Added map-image layer support to snapshot builder

* Added map-image support to participant component

* Added support for raster layers in legend

Raster layers contain their preview in a 'src' property instead of an appendable Node which was causing errors on trying to add a string instead.

PR cleanup.

* Added map-image sublayer opacity control to legend item

Layers listed from map image layer that have a `src` image now also have applied opacity to match the rendered raster on the map.

* Reverted map-image layer opacity control

The control has handed off to the children individually but went back to let the parent handle opacity for all children. I experienced and issue with that behavior before but it caused a bug where trying to apply an opacity to sub-layers that was higher than they were currently set did not apply the changes to the map raster itself until the page was reloaded.

Example:

Initial opacity for all sub-layers = .5;

Setting the opacity to 1 on all sub-layers did not change the raster on the map. However, upon refreshing the page, the change to opacity 1 was visible.

* Reversing the layer add o participant component

The layers were in the correct order in the snapshot builder but not in the participant view.