Fixed scenario builder default zoom
When the scenario was saved with a default zoom of '', it was casted to a 0 which always led to a world view zoom despite the default value being changed in the front-end application. Made the default zoom value in a scenario to be null so that the default zoom can be handled properly in the front-end.