fix(cpa): Short fixes for cpa nest and angular (#132)
* Fixed participants view not loading remote group layers

Group layers could only be built by explicitly providing the sub layers but there are certain layers that get casted from a service endpoint. Those group layers are now supported as well.

* Fixed workshop delete error

Fixed OMR relationship to cascade row deletions. This fixes the issue where workshops with contextual snapshots are cannot be removed.

* Fixed scenario builder default zoom

When the scenario was saved with a default zoom of '', it was casted to a 0 which always led to a world view zoom despite the default value being changed in the front-end application. Made the default zoom value in a scenario to be null so that the default zoom can be handled properly in the front-end.