Fixed result query, filtering, and stats
Split the where and filter clause generation because valve fetching and stat fetching needed different parts of the generated SQL to correctly display valves according to the search terms and any activated filters, meanwhile the valves stats only depended on the search term and not the activated filters. This fixed an issue where the valve stats were partial until filters were toggled.