feat(ues-dispatch): Dispatch valve intervention form (#139)
* Added group mapping to idp client

- Added jwt-decoder package
- Added ability to provide roles dictionary to the client and return a mapped role array for the roles associated with a user.

* Updated all affected apps' deployment pacakge.json

Added jwt-decoder to dependencies.

Removed extra package.json file for ues recycling app.

* Added intervention form

Needs to talk to server

* Added intervention service

* Attempt to get form submitting. Run into CORS issue.

Cannot submit from the front-end. Will probably need to move the intervention submission implementation to the backend to proxy the requests.

* Added cold water nest base lib

* Proxing all resource api requests to fix CORS issue

All the logic that was in the interventions service on the front-end has been moved the the backend which the front-end now calls.

Installed an additional package: form-data. Node doesn't have this API available.

* Added intervention events table in valve details view

* Provide current valve state in intervention event view

* Intervention and valve state parallel submissions

Logic that updated the valve state has been merged with the intervention form submission request.

* Fixed intervention new/edit views

Fixed issue where valve or intervention objects were not being resolved depending if the view was to edit or create a new intervention.

* Added form-level user feature restrictions

Restricted features based on feature access at the form level. Need to add checks on the back-end

* Reworked UES auth with Azure IDP

Fixes the issue where trying to access the application after the access token expires redirects to a "Not authorized" JSON error message.

* Fixed expires_at not availble when not logged in

* Ordering cold water details property list

Provide a priority list which is used to order those properties to the very top while the rest are unaffected.

This is to address OBJECTID and other properties needing to be at the top of the properties list preview.

* Adding guards to intervention posting

* Added required groups guard logic

Prevents users from accessing resources if they don't have the necessary assigned roles

* Implemented intervention event updating

* Added intervention delete support

Updated all feedback messages

* Added base feature data lib

Includes base attribute-table component. This will be a component that accepts a graphic or graphic attributes and renders them into a table

* Created attribute table component

Replaced valves property list with the attribute table.

* Added alias attribute table component

Similar to the attribute table component but instead renders the attribute aliases and field coded values.

Also adds two new pipes to this library: attribute-field pipe and field-coded-value pipe responsible for matching and returning a matching field object from a field name, and finding the named value from a field coded value.

* Fixed breadcrumb links in details and intervention views

* Refactoring CWV service in prep. for SS search

Added client-side optimizations for calculating and redering CW statistics as well as generalized queries for upcoming server-side paginated search result rendering.

* Added valves list search functionality

Searching input submits server side queries to accelerate app time to load.

Still need to clean up the filtering and categorization. There are some remnants of that logic still.

Also need to add result pagination

* Added search input clearing action

Slight alt text formatting for the valve ratio bar chart

* Removed list client-side categorizaton and filtering

* Fixed result query, filtering, and stats

Split the where and filter clause generation because valve fetching and stat fetching needed different parts of the generated SQL to correctly display valves according to the search terms and any activated filters, meanwhile the valves stats only depended on the search term and not the activated filters. This fixed an issue where the valve stats were partial until filters were toggled.

* Fixed valve fetching due to incorrect where clause buildling

* Cleaned up valve stat fetch repeated code

Condensed the if statements using some ternary function and string concatenation

* Fixed ratio bar when dividing by zero

Caused the bar to not update when query returned zero matching results.

* Added slight ratio bar width fill transition

* Review fixes