feat(ues-operations): Auth prompt routes (#149)
* Moved the re-directing logic from interceptor to auth svc

The auth service can then handle re-direction without the need of the interceptor for applications that don't have the app-wide http interceptor. Use cases are apps that want to display a prompt a session-expired or unauthenticated prompt/instruction instead of being immediately redirected when an auth request fails with 401/403.

* Common ngx auth guard has abiilty to return url tree

The underlying service call now returns the status code and on failed requests has the abiity to return a url tree constructed from provided route data. This makes it possible for the canActivate method to redirect to an unauthorized/session expired page or some other intermediary step.

* Added auth route module with components for auth redirects

Currently only a base session-expired component that will be auto-navigated to by the auth guard.

* Working session expired route/component

Implemented only in UES operations. When authentication fails, the app will redirect to an expired sesssion page with a timed redirect.