ci: Added a loop and event scenario if PM2 crashes to startup.bat file (#144)
Added a loop that is set on five minute intervals to the startup batch file for the PM2 window service that allows for NSSM to monitor whether or not PM2 has crashed. If PM2 has crashed it will attempt to resurrect and will publish an event in the logs entitled "nssmCheck".