Layer list supports layer descriptions
Descrpitions are triggered via a tooltip. The descriptions themselves are abusing the arcgis js API a *wee bit* since I am relying on setting a custom non-standard property to record and access the description.

Need to adjust the tooltip fixed positioning strategy to consider more viewport limits as the tooltips are currently overflowing off screen.

Drawer needed a change in the default transform value due to a misunderstanding of CSS's fixed positioning.