Moved device_id outside provider base entity
This resulted in the vehicle entity's `device_id` property not taking up as a primary column because overwritting decorators from an inherited class is not a thing.

This had the side effect of no vehicles column being updated and instead fresh rows were inserted on every vehicle scrape which filled up the database incredibly quickly.