Changed the workshop routing strategy
Used url params to set workshop and active timeline event. However, because the sidebar is now in charge of rendering all the viewer components, it needs to be able to toggle layers by a route. Since the particpant component will be destroyed on a route change, the state management needs to be moved outside of it so that there aren't any side effects when re-visiting the component.