Modified popup service to consider only valid graphics from a hit test snapshot
An issue came up where for the drawn features, the API was layering an additional graphic for a bounding box. This resulted in the popup service to be unable to determine the popup of the top-most graphic because it was not a graphic that contained a poupComponent assignment.

The popup service now only considers features for which the parent layer has a popupComponent assignment.

There was an attempt to preserve all other functionality.

Also fixed a couple of bugs that kept throwing error spam in the console.