Updated angular app pwa manifest for dev and prod
Couldn't figure out how to create multiple versions of the ngsw-config without it not releasing the webmanifest as a dist artifact. Angular is very specific aobut what the file names should be otherwise it errors out. This requires using JSON value substitution which I don't want to deal with right now.