Fix location status equality check
Some shapes were being resolved to falsy states even though the shape was added to the map. Came down to graphic attribute id prefixes based on geometry type as an attempt to differentiate them. Not needed for this purpose, so just went the lazy way and popping the marker id off the prefix.