Update asset => speaker, org, sponsor relationship
I had initially wanted to get away with many one-to-one relationships but that obviously causes issues with foreign key constraints. Somehow in local env, it allowed me to bypass this after a couple restarts. Not ideal since dev and prod wont have that luxury. Now asset entity has a many-to-one relationship with orgs, speakers, and sponsors. Makes it annoying when templating on the front-end but this is the way it has to be