Enable move-in-out service filtering
Residence filtering working only so far

Partial working for parking lots. There is a discreptancy between the lots in the moveinout service layer and the moveindays table (i.e. not all of the lots in the moveindays have a matched feature in the moveinout service, which results in only a small subset of lots being intersected and rendered