Potential fix for benefit array insertion index
There is a possible issue with new benefit insertion where its not inserted to the end of the array for some reason, despite the fact using .push. Using instead insert with an index refereence to hopefully address this issue. So far no issues, but unsure if unable to replicate a weird quirk or actually fixed