Update most relative router navigation
Most linking to and from application views were broken because they were inteded to be used in a root configuration, instead of a sub-module configuration. The fix required making navigation routing relative to a given route, often the route in which the component is loaded or the parent

Also fixes additional references to env service pulling from env configuration which are no longer present as all event configs have been consolidated.