Support default layer overrides
Given that events are now part of Aggiemap, it was not possible to alter the default properties for the default base layers. Specifically, certain events don't need to display things like construction zones in order to reduce visual noise. This creates a new layer source store which events service can call to set specific overrides based on the called even definition