Refactor application notification store
Current notifications store generation is 3. Migrations have been rewriten to simply clean slate if using gen 1 or 2 stores.

Rename EmittedNotification to PlatformNotification and renamed the nested notification props key. This to remove the need for a union type and supporting two types of notification models.

Ensured backwards-support for appliations that define their own notification presets. This is achieved by casting the notificationproperties into a platformnotification